slightly modified support for PFLAGNUM_5
git-svn-id: http://miranda.googlecode.com/svn/trunk@4666 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/skin/skinicons.c
@@ -227,7 +227,7 @@ static int LoadSkinProtoIcon(WPARAM wParam,LPARAM lParam)
 
 				if ( i < SIZEOF( statusIcons )) {
 					mir_snprintf( szFullPath, SIZEOF(szFullPath), "%s\\Plugins\\%s.dll", szPath, szProto );
-					if ( ExtractIconExA( szFullPath, statusIcons[i].resource_id, NULL, &hIcon, 1 ) > 0 ) {
+				if (( int )ExtractIconExA( szFullPath, statusIcons[i].resource_id, NULL, &hIcon, 1 ) > 0 ) {
 						DestroyIcon( hIcon );
 						sid.pszDefaultFile = szFullPath;
 						hIcon = NULL;
